% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bat_wav.R
\name{bat_wav}
\alias{bat_wav}
\title{create a batWav object}
\usage{
bat_wav(filename, channel = c("left", "right"), te.factor = 1, sample.rate,
  values)
}
\arguments{
\item{filename}{The name of the file}

\item{channel}{Select the left or the right channel}

\item{te.factor}{The factor to which the original sound was slowed down prior
to recording}

\item{sample.rate}{the realtime sample rate}

\item{values}{the recorded values}
}
\description{
create a batWav object
}
