% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batContourMeta-class.R
\docType{class}
\name{batContourMeta-class}
\alias{batContourMeta-class}
\title{The batContourMeta class}
\description{
It holds metadata on bat recordings
}
\section{Slots}{

  \describe{
   \item{\code{Recording}}{A data.frame with ID, Fingerprint, Filename, Timestamp, SampleRate, SampleRate, TEFactor, LeftChannel}
   \item{\code{Spectrogram}}{A data.frame with ID, Fingerprint, WindowMS, WindowN, Overlap, Recording}
   \item{\code{Pulse}}{A data.frame with ID, Fingerprint, Spectrogram, PeakX, PeakY, PeakAmplitude}
   \item{\code{Contour}}{A data.frame with ID, Fingerprint, Pulse, ContourAmplitude, ContourStep, ContourMax}
  }
}

