% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batPulse-class.R
\docType{class}
\name{batPulse-class}
\alias{batPulse-class}
\title{The batPulse class}
\description{
Holds the relevant content of wav files
}
\section{Slots}{

  \describe{
   \item{\code{Plot}}{A SpatialPolygonsDataFrame with the contours}
  }
}

