% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batpulse2db.R
\name{batpulse2db}
\alias{batpulse2db}
\title{Store a batPulse object into a database}
\usage{
batpulse2db(x, connection)
}
\arguments{
\item{x}{a batPulse object}

\item{connection}{a DBI connection to the database}
}
\description{
Store a batPulse object into a database
}
