% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batpulse_rda2list.R
\name{batpulse_rda2list}
\alias{batpulse_rda2list}
\title{Convert batPulse objects stored in files to a list}
\usage{
batpulse_rda2list(files)
}
\arguments{
\item{files}{A vector of filenames}
}
\description{
Convert batPulse objects stored in files to a list
}

