% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batPulseParameter-class.R
\docType{class}
\name{batPulseParameter-class}
\alias{batPulseParameter-class}
\title{The batPulseParameter class}
\description{
Holds the relevant content of wav files
}
\section{Slots}{

  \describe{
   \item{\code{Recording}}{The metadata on the recording}
   \item{\code{SpecGram}}{The actual spectrogram}
  }
}

