% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batRecording-class.R
\docType{class}
\name{batRecording-class}
\alias{batRecording-class}
\title{The batRecording class}
\description{
It holds metadata on bat recordings
}
\section{Slots}{

  \describe{
   \item{\code{Recording}}{A data.frame with ID, Fingerprint, Filename, Timestamp, SampleRate, SampleRate, TEFactor, LeftChannel}
  }
}

