% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batSpectrogram-class.R
\docType{class}
\name{batSpectrogram-class}
\alias{batSpectrogram-class}
\title{The batSpectrogram class}
\description{
Holds the relevant content of wav files
}
\section{Slots}{

  \describe{
   \item{\code{Recording}}{The metadata on the recording}
   \item{\code{SpecGram}}{The actual spectrogram}
  }
}

