% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batSpectrogramMeta-class.R
\docType{class}
\name{batSpectrogramMeta-class}
\alias{batSpectrogramMeta-class}
\title{The batSpectrogramMeta class}
\description{
It holds metadata on bat recordings
}
\section{Slots}{

  \describe{
   \item{\code{Recording}}{A data.frame with ID, Fingerprint, Filename, Timestamp, SampleRate, SampleRate, TEFactor, LeftChannel}
   \item{\code{Spectrogram}}{A data.frame with ID, Fingerprint, WindowMS, WindowN, Overlap, Recording}
  }
}

