% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batWav-class.R
\docType{class}
\name{batWav-class}
\alias{batWav-class}
\title{The batWav class}
\description{
Holds the relevant content of wav files
}
\section{Slots}{

  \describe{
   \item{\code{SampleRate}}{The realtime sample rate (i.e. recording sample rate x time expantion factor)}
   \item{\code{TEFactor}}{Time expantion factor}
   \item{\code{Value}}{The actual recorded values}
   \item{\code{Fingerprint}}{The fingerprint of the object}
  }
}

