% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batWavMeta-class.R
\docType{class}
\name{batWavMeta-class}
\alias{batWavMeta-class}
\title{The batWavMeta class}
\description{
It holds metadata on bat recordings
}
\section{Slots}{

  \describe{
   \item{\code{Recording}}{A data.frame with ID, Fingerprint, Filename, Timestamp, SampleRate, SampleRate, TEFactor, LeftChannel}
  }
}

