% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_box.R
\name{calculate_box}
\alias{calculate_box}
\title{Calcalute the box around a pulse}
\usage{
calculate_box(pulses)
}
\arguments{
\item{pulses}{a batPulse object}
}
\description{
Calcalute the box around a pulse
}

