% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fingerprint.R
\name{check_fingerprint}
\alias{check_fingerprint}
\title{Check the matching id for a fingerprint}
\usage{
check_fingerprint(src, table, current)
}
\arguments{
\item{src}{a scr_dbi object}

\item{table}{the name of the table in the database}

\item{current}{the dataframe to upload}
}
\description{
Check the matching id for a fingerprint
}
