% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/combine_rda.R
\name{combine_rda}
\alias{combine_rda}
\title{combine the data from several rda files into a object}
\usage{
combine_rda(path, keep.spectrogram = FALSE)
}
\arguments{
\item{path}{the path in which to look recursively}

\item{keep.spectrogram}{TRUE keeps the spectrograms, FALSE discards them}
}
\description{
combine the data from several rda files into a object
}

