% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_rds.R
\name{combine_rds}
\alias{combine_rds}
\title{combine the data from several rda files into a object}
\usage{
combine_rds(path, keep.spectrogram = FALSE)
}
\arguments{
\item{path}{the path in which to look recursively}

\item{keep.spectrogram}{TRUE keeps the spectrograms, FALSE discards them}
}
\description{
combine the data from several rda files into a object
}

