% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_db.R
\name{connect_db}
\alias{connect_db}
\title{Connect to a rhinolophus database}
\usage{
connect_db(path)
}
\arguments{
\item{path}{the location of the database}
}
\description{
Connect to a rhinolophus database
}
