% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/contour_2_pulse.R
\name{contour2pulse}
\alias{contour2pulse}
\title{Cluster contours into pulses}
\usage{
contour2pulse(contours, min.peak = 20)
}
\arguments{
\item{contours}{A SpatialPolygonsDataFrame with contours. E.g. the output of \code{\link{spectrogram.contour}}}

\item{min.peak}{the minimum height of the highest contour of a peak}
}
\description{
Cluster contours into pulses
}

