% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_timestamp.R
\name{correct_timestamp}
\alias{correct_timestamp}
\title{Shift the timestamp 2 hours back}
\usage{
correct_timestamp(path)
}
\arguments{
\item{path}{the root folder to be changed}
}
\description{
Note that ALL files in the path will be recursively updated
}
