% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elliptic_fourier.R
\name{elliptic_fourier}
\alias{elliptic_fourier}
\title{Calculate the elliptic fourier transformation}
\usage{
elliptic_fourier(x, f_max = 30)
}
\arguments{
\item{x}{a data.frame containing X1, X2 and row_id}

\item{f_max}{the maximum cyclic frequency}
}
\description{
Calculate the elliptic fourier transformation
}
