% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_predict.R
\name{ensemble.predict}
\alias{ensemble.predict}
\title{Get predicted probabilities of class membership based on an ensemble of Kohonen self organizing maps}
\usage{
ensemble.predict(truth, pattern, newdata, n.ensemble = 10)
}
\arguments{
\item{truth}{the known classification}

\item{pattern}{the patterns of the known classification}

\item{newdata}{the pattern to get predictions for}

\item{n.ensemble}{Number of kohonen SOM's to use}
}
\description{
Get predicted probabilities of class membership based on an ensemble of Kohonen self organizing maps
}
