% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_parameter.R
\name{extract_parameter}
\alias{extract_parameter}
\title{Extract the coefficients of the pattern from a batPulse object}
\usage{
extract_parameter(object, n.part = 5)
}
\arguments{
\item{object}{a batPulse object}

\item{n.part}{the number of Fourier dimensions}
}
\description{
Extract the coefficients of the pattern from a batPulse object
}

