% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_parameter.R
\name{extract_pulse}
\alias{extract_pulse}
\title{Extract the Fourier components of a pulse}
\usage{
extract_pulse(fingerprint, object, n.part)
}
\arguments{
\item{fingerprint}{the fingerprint of the pulse}

\item{object}{the batPulse object}

\item{n.part}{the number of Fourier components}
}
\description{
Extract the Fourier components of a pulse
}

