% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fft_pulse.R
\name{fft.pulse}
\alias{fft.pulse}
\title{Calculate the fast fourier transformation for each pulse}
\usage{
fft.pulse(pulses, spectrogram, n.fourier = 30)
}
\arguments{
\item{pulses}{A list of pulses. E.g. the output of \code{\link{contour2pulse}}}

\item{spectrogram}{The spectrogram. E.g. the output of \code{\link{wav.2.spectrogram}}}

\item{n.fourier}{The number of required Fourier components in each direction.}
}
\description{
Calculate the fast fourier transformation for each pulse
}

