% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/find_pulses.R
\name{find.pulses}
\alias{find.pulses}
\title{Find the location of possible pulses in a spectrogram}
\usage{
find.pulses(spectrogram, min.contour = 10, min.peak = 20)
}
\arguments{
\item{spectrogram}{The spectrogram}

\item{min.contour}{The minimum amplitude of the pulse}

\item{min.peak}{Take only contour into account with a maximum amplitude of at least min.peak}
}
\description{
Find the location of possible pulses in a spectrogram
}

