% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_kohonen.R
\name{fit_kohonen}
\alias{fit_kohonen}
\title{Get validation summary for raw data}
\usage{
fit_kohonen(path)
}
\arguments{
\item{path}{the path of the _truth.rds and _parameter.rds files}
}
\description{
Get validation summary for raw data
}

