% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier_ellipse.R
\name{fourier_ellipse}
\alias{fourier_ellipse}
\title{calculate the ellipse parameters from a Fourier parameters}
\usage{
fourier_ellipse(fourier_param)
}
\arguments{
\item{fourier_param}{a 2x2 matrix with Fourier parameters}
}
\value{
a vector with 3 values: the mayor semi-axis, the minor semi-axis and the rotation (in radians)
}
\description{
calculate the ellipse parameters from a Fourier parameters
}
