% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_activity_list.R
\name{get_activity_list}
\alias{get_activity_list}
\title{get the acitivity list from the database}
\usage{
get_activity_list(connection)
}
\arguments{
\item{connection}{the connection to the database}
}
\description{
get the acitivity list from the database
}
