% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contours.R
\name{get_contours}
\alias{get_contours}
\title{get contours as SpatialPolygons}
\usage{
get_contours(x, levels)
}
\arguments{
\item{x}{a raster}

\item{levels}{a vector with the required levels}
}
\description{
get contours as SpatialPolygons
}
