% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species.R
\name{get_species}
\alias{get_species}
\title{get the species information of all recordings in a directory}
\usage{
get_species(path, write_csv = TRUE)
}
\description{
get the species information of all recordings in a directory
}
