% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_list.R
\name{get_species_list}
\alias{get_species_list}
\title{get the species list from the database}
\usage{
get_species_list(connection)
}
\arguments{
\item{connection}{the connection to the database}
}
\description{
get the species list from the database
}
