% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulse_matrix.R
\name{interpolate_pulse}
\alias{interpolate_pulse}
\title{Convert a single pulse to a sparse matrix}
\usage{
interpolate_pulse(spectrogram.raster, this.box)
}
\arguments{
\item{spectrogram.raster}{a spectrogram raster}

\item{this.box}{the bounding box of the pulse}
}
\description{
Convert a single pulse to a sparse matrix
}

