% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxel_pulse.R
\name{interpolate_voxel}
\alias{interpolate_voxel}
\title{Calculate the voxel for an individual pulse}
\usage{
interpolate_voxel(spectrogram.raster, this.box)
}
\arguments{
\item{spectrogram.raster}{a spectrogram raster}

\item{this.box}{the bounding box of the pulse}
}
\description{
Calculate the voxel for an individual pulse
}

