% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_dl.R
\name{label_dl}
\alias{label_dl}
\title{Retrieve a random recording id}
\usage{
label_dl(connection)
}
\arguments{
\item{connection}{a connection to the database}
}
\description{
Retrieve a random recording id
}
