% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_unsupervised.R
\name{label_unsupervised}
\alias{label_unsupervised}
\title{Retrieve the recording id which contains unlabeled pulses for many classes}
\usage{
label_unsupervised(connection)
}
\arguments{
\item{connection}{a connection to the database}
}
\description{
Retrieve the recording id which contains unlabeled pulses for many classes
}
