% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/native_dl.R
\name{naive_dl}
\alias{naive_dl}
\title{Create prediction with a naive deep learning model}
\usage{
naive_dl(path)
}
\arguments{
\item{path}{the location of the database}
}
\description{
Create prediction with a naive deep learning model
}
