% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/next_power_2.R
\name{next_power_2}
\alias{next_power_2}
\title{Calculate the next power of 2}
\usage{
next_power_2(x)
}
\arguments{
\item{x}{the number of which to calculate the next power of 2}
}
\description{
Calculate the next power of 2
}

