% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotpattern.R
\name{plotpattern}
\alias{plotpattern}
\title{Plot a single pattern}
\usage{
plotpattern(plotdata, amplitudes, times, frequencies, reference = c(30, 40,
  45, 50, 55) * 1000)
}
\arguments{
\item{plotdata}{The output of prepareplot}

\item{amplitudes}{Limits of amplitudes. Lower amplitudes will be gray, higer amplitudes will be white.}

\item{times}{Limits of time}

\item{frequencies}{Limits of frequencies}

\item{reference}{Frequecies of reference lines to plot}
}
\description{
Plot a single pattern
}
