% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_kohonen.R
\name{predict_kohonen}
\alias{predict_kohonen}
\title{Predict the classification according to a set of supervised Kohonen's self-organising maps}
\usage{
predict_kohonen(models, dataset_matrix)
}
\arguments{
\item{models}{a list of supervised Kohonen's self-organising maps}

\item{dataset_matrix}{a matrix with the parameter values of the new observations}
}
\description{
Predict the classification according to a set of supervised Kohonen's self-organising maps
}

