% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_analysis.R
\name{prepare_analysis}
\alias{prepare_analysis}
\title{Prepare all pulse parameters for analysis}
\usage{
prepare_analysis(path, n.part = 30)
}
\arguments{
\item{path}{the path of the pulses}

\item{n.part}{the number of Fourier dimensions}
}
\description{
Prepare all pulse parameters for analysis
}

