% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareplot.R
\name{prepareplot}
\alias{prepareplot}
\title{Prepare the plot a single pattern}
\usage{
prepareplot(pattern)
}
\arguments{
\item{pattern}{A single pattern}
}
\description{
Prepare the plot a single pattern
}
