% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pulse_border.R
\name{pulse_border}
\alias{pulse_border}
\title{Calculate the bounding box of a pulse}
\usage{
pulse_border(pulse)
}
\arguments{
\item{pulse}{a dataframe with a single pulse}
}
\description{
Calculate the bounding box of a pulse
}

