% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulse_label.R
\name{pulse_label}
\alias{pulse_label}
\title{Generate labels for the pulse}
\usage{
pulse_label(borders)
}
\arguments{
\item{borders}{a SpatialPolygonsDataFrame object}
}
\description{
Generate labels for the pulse
}

