% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulse_matrix.R
\name{pulse_matrix}
\alias{pulse_matrix}
\title{Extract the parameters of all pulses as a sparse matrix}
\usage{
pulse_matrix(pulses)
}
\arguments{
\item{pulses}{a batPulse object}
}
\description{
Extract the parameters of all pulses as a sparse matrix
}

