% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulse_parameter.R
\name{pulse_parameter}
\alias{pulse_parameter}
\title{Convert pulses into parameters}
\usage{
pulse_parameter(pulses)
}
\arguments{
\item{pulses}{a SpatialPolygonsDataFrame with pulses}
}
\description{
Convert pulses into parameters
}
