% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda2pattern.R
\name{rda2pattern}
\alias{rda2pattern}
\title{Convert rda files to a set of patterns}
\usage{
rda2pattern(path, n.part = 10)
}
\arguments{
\item{path}{The path of the rda files}

\item{n.part}{The number of fourier components to use. Defaults to 10}
}
\description{
Convert rda files to a set of patterns
}
