% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds2db.R
\name{rds2db}
\alias{rds2db}
\title{Store all information from individual rds files into a database}
\usage{
rds2db(path, db_path = path)
}
\arguments{
\item{path}{the path to the rds files}

\item{db_path}{the path to the database}
}
\description{
Store all information from individual rds files into a database
}
