% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wav.R
\name{read_wav}
\alias{read_wav}
\title{read a WAV file}
\usage{
read_wav(filename, channel = c("left", "right"), te.factor = 10,
  max.length = 1.8)
}
\arguments{
\item{filename}{The name of the file}

\item{channel}{Select the left or the right channel}

\item{te.factor}{The factor to which the original sound was slowed down prior
to recording}

\item{max.length}{The maximum length to read from the wav file. The time is
in original seconds. So te.factor = 10 and max.length = 1.8 will read the
first 10 * 1.8 = 18 seconds from the wav file.}
}
\description{
read a WAV file
}
\examples{
 wav = read_wav(
   system.file("demo_wav/leislers.wav", package = "rhinolophus")
 )
}

