% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_pulse.R
\name{reconstruct.pulse}
\alias{reconstruct.pulse}
\title{reconstruct a pulse}
\usage{
reconstruct.pulse(pulses, fingerprint, plot.it = TRUE)
}
\arguments{
\item{pulses}{The fourier representation of the pulse}
}
\description{
reconstruct a pulse
}

