% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct.R
\name{reconstruct}
\alias{reconstruct}
\title{Reconstruct pulse from their unscales Fourier parameters}
\usage{
reconstruct(params, n = 100)
}
\arguments{
\item{params}{the unscales Fourier parameters}

\item{n}{the number of point in each pulse}
}
\description{
Reconstruct pulse from their unscales Fourier parameters
}
