% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spectrogram_contour.R
\name{spectrogram.contour}
\alias{spectrogram.contour}
\title{Calculate contour polygons from a spectrogram}
\usage{
spectrogram.contour(spectrogram, contour.level = NULL)
}
\arguments{
\item{spectrogram}{A spectrogram as generated by \code{\link{wav_2_spectrogram}}}

\item{contour.level}{A vector of contour levels. If one number it is used as the step size in the vector.}
}
\description{
Calculate contour polygons from a spectrogram
}

