% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spectrogram_raster.R
\name{spectrogram_raster}
\alias{spectrogram_raster}
\title{Convert a spectrogram to a raster object}
\usage{
spectrogram_raster(spectrogram)
}
\arguments{
\item{spectrogram}{an object of the class specgram}
}
\description{
Convert a spectrogram to a raster object
}

