% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_unsupervised.R
\name{status_unsupervised}
\alias{status_unsupervised}
\title{Create an overview of the labels and unsupervised classes}
\usage{
status_unsupervised(connection)
}
\arguments{
\item{connection}{the connection to the database}
}
\description{
Create an overview of the labels and unsupervised classes
}
