% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_manual.R
\name{store_manual}
\alias{store_manual}
\title{write the manual classification to the database}
\usage{
store_manual(connection, manual)
}
\arguments{
\item{connection}{the connection to the database}

\item{manual}{a data.frame with manual classification}
}
\description{
write the manual classification to the database
}
