% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainrda.R
\name{trainrda}
\alias{trainrda}
\title{Train a Kohonen network}
\usage{
trainrda(pattern, truth)
}
\arguments{
\item{pattern}{A matrix with patterns}

\item{truth}{A data.frame with the truth}
}
\description{
Train a Kohonen network
}
