% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_truth.R
\name{validate_truth}
\alias{validate_truth}
\title{Get validation summary for raw data}
\usage{
validate_truth(path)
}
\arguments{
\item{path}{the path of the _truth.rds and _parameter.rds files}
}
\description{
Get validation summary for raw data
}

