% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxel_pulse.R
\name{voxel_pulse}
\alias{voxel_pulse}
\title{Calculate the voxels for all pulses in a spectrogram}
\usage{
voxel_pulse(pulses)
}
\arguments{
\item{pulses}{a batPulse object}
}
\description{
Calculate the voxels for all pulses in a spectrogram
}

