% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wav2rds.R
\name{wav2rds}
\alias{wav2rds}
\title{Find the pulses in a bunch of wav files and save them to rda files}
\usage{
wav2rds(path, te.factor = 10, channel = c("right", "left"),
  max.length = 1.8, window.ms = 2, min.peak = 40, min.amplitude = 10,
  delta.amplitude = 30, n.fourier = 30, action = c("append", "update",
  "overwrite"))
}
\arguments{
\item{path}{The path of the wav files. The function will reach recursive in
the subdirectories of the path.}

\item{te.factor}{The factor to which the original sound was slowed down prior
to recording}

\item{channel}{Select the left or the right channel}

\item{max.length}{The maximum length to read from the wav file. The time is
in original seconds. So te.factor = 10 and max.length = 1.8 will read the
first 10 * 1.8 = 18 seconds from the wav file.}

\item{window.ms}{The size of the window in microseconds. Default to 1.}

\item{min.peak}{Take only pulses into account with a maximum amplitude of at least min.peak}

\item{min.amplitude}{The minimum amplitude of the pulse}

\item{delta.amplitude}{The difference in amplitude between the peak of the pulse and the lowest value in the pulse.}

\item{n.fourier}{The number of required Fourier components in each direction.}

\item{action}{What to do when a rds file exists? \code{"append"} will add only new information. Existing information will be intact. \code{"update"} will keep all information and update existing information. \code{"overwrite"} remove the rds file and creates a new one. Defaults to \code{"append"}.}
}
\description{
Find the pulses in a bunch of wav files and save them to rda files
}

