% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wav2spectrogram.R
\name{wav2spectrogram}
\alias{wav2spectrogram}
\title{Convert the output of read_wav to a spectrogram}
\usage{
wav2spectrogram(wav, window.ms = 1)
}
\arguments{
\item{wav}{An object as created by \code{\link{read_wav}}}

\item{window.ms}{The size of the window in microseconds. Default to 1.}
}
\description{
Convert the output of read_wav to a spectrogram
}
\examples{
 wav2spectrogram(
   wav = read_wav(
     system.file("demo_wav/leislers.wav", package = "rhinolophus")
   )
 )
}
