% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bat_som.R
\name{bat_som}
\alias{bat_som}
\title{Calculate a self organizing map on bat pulse features}
\usage{
bat_som(x, dims = rep(floor(log(nrow(x))), 2), topo = c("hexagonal",
  "rectangular"))
}
\arguments{
\item{x}{the output of \code{db2ml()}}

\item{dims}{dimensions of the grid}

\item{topo}{hexagonal or rectangular topology}
}
\description{
Calculate a self organizing map on bat pulse features
}
