% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batWav-class.R
\docType{class}
\name{batWav-class}
\alias{batWav-class}
\title{The batWav class}
\description{
Holds the relevant content of wav files
}
\section{Slots}{

  \describe{
   \item{\code{Recording}}{The metadata on the recording}
   \item{\code{Values}}{The actual recorded values}
  }
}

