% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour2pulse.R
\name{contour2pulse}
\alias{contour2pulse}
\title{Cluster contours into pulses}
\usage{
contour2pulse(contours, min.peak = 20)
}
\arguments{
\item{contours}{A SpatialPolygonsDataFrame with contours. E.g. the output of \code{\link{spectrogram_contour}}}

\item{min.peak}{the minimum height of the highest contour of a peak}
}
\description{
Cluster contours into pulses
}
\examples{
\dontrun{
wav <- read_wav(
  system.file("demo_wav/leislers.wav", package = "rhinolophus")
)
spectrogram <- wav2spectrogram(wav)
contours <- spectrogram_contour(spectrogram)
pulses <- contour2pulse(contours)
plot(pulses[[1]], asp = 0.25)
axis(1)
axis(2)
}
}
