% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db2ml.R
\name{db2ml}
\alias{db2ml}
\title{Extract a machine learning dataset from the database}
\usage{
db2ml(path, n.harmonic = 30, contour.amplitude = -25)
}
\arguments{
\item{path}{the path of the rhinolophus database}

\item{n.harmonic}{the number of harmonics to extract}

\item{contour.amplitude}{the amplitude of the contour to extract}
}
\description{
Extract a machine learning dataset from the database
}
