% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db2sp.R
\name{db2sp}
\alias{db2sp}
\title{convert the contours of a given recording in the database into a SpatialPolygonsDataFrame}
\usage{
db2sp(connection, recording, dl = FALSE)
}
\arguments{
\item{connection}{the connection to the database}

\item{recording}{the id of the recording}

\item{dl}{get the naive deep learning classification}
}
\description{
convert the contours of a given recording in the database into a SpatialPolygonsDataFrame
}
