% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pulse.R
\name{extract_pulse}
\alias{extract_pulse}
\title{Extract pulses from a spectrogram}
\usage{
extract_pulse(spectrogram, min.peak = 30, contour.step = 10)
}
\arguments{
\item{spectrogram}{a spectrogram}

\item{min.peak}{minimum amplitude of the peak. Defaults to 30 dB.}

\item{contour.step}{step size of the contour. Defaults to 10 dB.}
}
\description{
Extract pulses from a spectrogram
}
