% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/find_pulses.R
\name{find.pulses}
\alias{find.pulses}
\title{Find the location of possible pulses in a spectrogram}
\usage{
find.pulses(spectrogram, min.kHz = 10, local.peak.ms = 10)
}
\arguments{
\item{spectrogram}{The spectrogram}

\item{min.kHz}{Ignore spectogram data below this frequency}

\item{local.peak.ms}{The time (in milliseconds) over which the current amplitude must be equal to the local maximum or minimum}
}
\description{
Find the location of possible pulses in a spectrogram
}

