% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_pulses.R
\name{find_pulses}
\alias{find_pulses}
\title{Find the location of possible pulses in a spectrogram}
\usage{
find_pulses(spectrogram, min.peak = 40, min.amplitude = 10,
  delta.amplitude = 10)
}
\arguments{
\item{spectrogram}{The spectrogram}

\item{min.peak}{Take only pulses into account with a maximum amplitude of at least min.peak}

\item{min.amplitude}{The minimum amplitude of the pulse}

\item{delta.amplitude}{The difference in amplitude between the peak of the pulse and the lowest value in the pulse.}
}
\description{
Find the location of possible pulses in a spectrogram
}
\examples{
 wav <- read_wav(
   system.file("demo_wav/leislers.wav", package = "rhinolophus")
 )
 spectrogram <- wav2spectrogram(wav)
 find_pulses(spectrogram)
}

