% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft_pulse.R
\name{interpolate_fft}
\alias{interpolate_fft}
\title{Calculate the fast Fourier transformation of an indivual pulse}
\usage{
interpolate_fft(spectrogram.raster, this.box, n.fourier)
}
\description{
Calculate the fast Fourier transformation of an indivual pulse
}

