% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxel_pulse.R
\name{interpolate_voxel}
\alias{interpolate_voxel}
\title{Calculate the voxel for an individual pulse}
\usage{
interpolate_voxel(spectrogram.raster, this.box)
}
\description{
Calculate the voxel for an individual pulse
}

