% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midpoint_ramp.R
\name{midpoint_ramp}
\alias{midpoint_ramp}
\title{Generate a colour ramp true a midpoint}
\usage{
midpoint_ramp(x, mid.point = 10, n = 20)
}
\arguments{
\item{x}{a RasterLayer object}

\item{mid.point}{the midpoint of the colour ramp}

\item{n}{the approximate number of breaks}
}
\description{
Generate a colour ramp true a midpoint
}

