% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next_power_2.R
\name{next_power_2}
\alias{next_power_2}
\title{Calculate the next power of 2}
\usage{
next_power_2(x)
}
\arguments{
\item{x}{the number of which to calculate the next power of 2}
}
\description{
Calculate the next power of 2
}
\examples{
next_power_2(3)
next_power_2(12345)
next_power_2(c(7, 9))
}
