% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_pulse.R
\name{reconstruct.pulse}
\alias{reconstruct.pulse}
\title{reconstruct a pulse}
\usage{
reconstruct.pulse(pulses, fingerprint)
}
\arguments{
\item{pulse.fft}{The fourier representation of the pulse}

\item{local.time}{Should the pulse be centered (TRUE) or display the original timing (FALSE)}

\item{plot.it}{Should the reconstructed pulse be plotted?}

\item{asp}{The aspect ratio}
}
\description{
reconstruct a pulse
}

