% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sound_png.R
\name{sound_png}
\alias{sound_png}
\title{Store sounds from the spectrogram as png}
\usage{
sound_png(sp, min.amp = 30, soundbite = 0.01, path, verbose = TRUE)
}
\arguments{
\item{sp}{a batSpectgram object}

\item{min.amp}{the minimal amplitude of the strongest part of the sound in dB}

\item{soundbite}{the maximal length of the sound in seconds}

\item{path}{the path to store the png and metadata files}

\item{verbose}{display the name of the current wav file}
}
\description{
Store sounds from the spectrogram as png
}
