% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised.R
\name{unsupervised}
\alias{unsupervised}
\title{Create a new unsupervised classification}
\usage{
unsupervised(path, n.harmonic = 10, contour.amplitude = -25, class = NULL,
  dims = c(4, 4), topo = c("rectangular", "hexagonal"), ...)
}
\arguments{
\item{path}{the path of the rhinolophus database}

\item{n.harmonic}{the number of harmonics to extract}

\item{contour.amplitude}{the amplitude of the contour to extract}

\item{dims}{dimensions of the grid}

\item{topo}{rectangular or hexagonal topology}

\item{...}{arguments past to \code{\link[kohonen]{som}}}
}
\description{
Create a new unsupervised classification
}
