% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wav2rds.R
\name{wav2rds}
\alias{wav2rds}
\title{Find the pulses in a bunch of wav files and save them to rda files}
\usage{
wav2rds(path, te.factor = 10, channel = c("right", "left"),
  max.length = 1.8, window.ms = 2, min.peak = 20, min.amplitude = 10,
  delta.amplitude = c(10, 20), n.fourier = 30, overwrite = FALSE)
}
\arguments{
\item{path}{The path of the wav files. The function will reach recursive in
the subdirectories of the path.}

\item{te.factor}{The time-expansion factor used in the wav files. Defaults
to 10.}

\item{channel}{Which channel will be read. Defaults to "right".}

\item{window.ms}{The size of the window for fourier transformation of the
sound in milliseconds. The time-expansion factor is taken into account.
Defaults to 1.}

\item{min.peak}{Take only pulses into account with a maximum amplitude of at least min.peak}

\item{min.amplitude}{The minimum amplitude of the pulse}

\item{delta.amplitude}{The difference in amplitude between the peak of the pulse and the lowest value in the pulse.}

\item{n.fourier}{The number of parameters of the Fourier Transformation to
store per dimension. Defaults to 30.}

\item{overwrite}{Overwrite existing rda files. Defaults to FALSE}
}
\description{
Find the pulses in a bunch of wav files and save them to rda files
}

