% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wav2rds.R
\name{wav2rds}
\alias{wav2rds}
\title{Extract pulses and store them in an rds file}
\usage{
wav2rds(path, te.factor = 1, channel = c("left", "right"), window.ms = 1,
  overlap = 0.9, min.peak = 30, contour.step = 10, contour.n = 3)
}
\arguments{
\item{path}{the path of the files}

\item{te.factor}{The factor to which the original sound was slowed down prior
to recording}

\item{channel}{Select the left or the right channel}

\item{window.ms}{The size of the window in microseconds. Default to 1.}

\item{overlap}{The overlap of two windows. Defaults to 0.9.}

\item{min.peak}{minimum amplitude of the peak. Defaults to 30 dB.}

\item{contour.step}{step size of the contour. Defaults to 10 dB.}

\item{contour.n}{the maximum number of contours per pulse. Defaults to 3.}
}
\description{
Extract pulses and store them in an rds file
}
